CLI commands added, API response parsed into List<Song>
