+++ pom.xml
@@ -18,6 +18,11 @@
 		<clojure.version>1.11.1</clojure.version>
     </properties>
     <dependencies>
+        <dependency>
+            <groupId>commons-cli</groupId>
+            <artifactId>commons-cli</artifactId>
+            <version>1.5.0</version>
+        </dependency>
         <!-- Logging -->
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -87,7 +92,7 @@
                             <transformers>
                                 <transformer
                                     implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.walit.streamline.Core</mainClass>
+                                    <mainClass>com.walit.streamline.Driver</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
