+++ src/main/java/com/walit/streamline/Utilities/Internal/Mode.java
@@ -1,0 +1,7 @@
+package com.walit.streamline.Utilities.Internal;
+
+public enum Mode {
+    TERMINAL, // DEFAULT BEHAVIOR
+    HEADLESS,
+    TESTING
+}
